+++ components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineSession.kt
@@ -1157,4 +1157,11 @@ abstract class EngineSession(
      * Gets the page text content of this session
      */
     open fun getPageContent(onResult: (String) -> Unit, onException: (Throwable) -> Unit) = Unit
+
+    /**
+     * Allow the Engine to handle back navigation events to dismiss some HTML elements such as &lt;dialog&gt;.
+     *
+     * @param onResult callback invoked if the engine API returned a valid response.
+     */
+    abstract fun processBackPressed(onResult: (Boolean) -> Unit)
 }