+++ components/support/test-fakes/src/main/java/mozilla/components/support/test/fakes/engine/TestEngineSession.kt
@@ -101,4 +101,6 @@ open class TestEngineSession(override val settings: Settings = DefaultSettings()
     override fun exitFullScreenMode() = Unit
 
     override fun purgeHistory() = Unit
+
+    override fun processBackPressed(onResult: (Boolean) -> Unit) = Unit
 }